[#9513] Move path related things in a ProcedurePathConcern
