+++ app/models/procedure.rb
@@ -9,6 +9,7 @@ class Procedure < ApplicationRecord
   include ProcedureSVASVRConcern
   include ProcedureChorusConcern
   include ProcedurePublishConcern
+  include ProcedurePathConcern
   include PiecesJointesListConcern
   include ColumnsConcern
 
@@ -232,7 +233,6 @@ class Procedure < ApplicationRecord
 
   validates :replaced_by_procedure_id, presence: true, if: :closing_reason_internal_procedure?
 
-  validates :path, presence: true, format: { with: /\A[a-z0-9_\-]{3,200}\z/ }, uniqueness: { scope: [:path, :closed_at, :hidden_at, :unpublished_at], case_sensitive: false }
   validates :duree_conservation_dossiers_dans_ds, allow_nil: false,
                                                   numericality: {
                                                     only_integer: true,
@@ -294,8 +294,6 @@ class Procedure < ApplicationRecord
   before_save :update_juridique_required
   after_save :extend_conservation_for_dossiers
 
-  after_initialize :ensure_path_exists
-  before_save :ensure_path_exists
   after_create :ensure_defaut_groupe_instructeur
 
   include AASM
@@ -329,30 +327,6 @@ class Procedure < ApplicationRecord
 
   def dossiers_close_to_expiration
     dossiers.close_to_expiration.count
-  end
-
-  def suggested_path(administrateur)
-    if path_customized?
-      return path
-    end
-    slug = libelle&.parameterize&.first(50)
-    suggestion = slug
-    counter = 1
-    while !path_available?(suggestion)
-      counter = counter + 1
-      suggestion = "#{slug}-#{counter}"
-    end
-    suggestion
-  end
-
-  def other_procedure_with_path(path)
-    Procedure.publiees
-      .where.not(id: self.id)
-      .find_by(path: path)
-  end
-
-  def path_available?(path)
-    other_procedure_with_path(path).blank?
   end
 
   def canonical_procedure_child?(procedure)
@@ -423,10 +397,6 @@ class Procedure < ApplicationRecord
 
   def feature_enabled?(feature)
     Flipper.enabled?(feature, self)
-  end
-
-  def path_customized?
-    !path.match?(/[[:xdigit:]]{8}-[[:xdigit:]]{4}-[[:xdigit:]]{4}-[[:xdigit:]]{4}-[[:xdigit:]]{12}/)
   end
 
   def organisation_name
@@ -799,12 +769,6 @@ class Procedure < ApplicationRecord
   def check_juridique
     if juridique_required? && (cadre_juridique.blank? && !deliberation.attached?)
       errors.add(:cadre_juridique, " : veuillez remplir le texte de loi ou la délibération")
-    end
-  end
-
-  def ensure_path_exists
-    if self.path.blank?
-      self.path = SecureRandom.uuid
     end
   end
